Revert "Feature/40 restrctr src pkg"
