+++ src/__init__.py
@@ -1,0 +1,4 @@
+from .olsmodel import OLSRegressor
+from .polymodel import PolynomialModel
+from .transformer_boxcox import BoxCoxTransformer
+from .image_dataset import ImageDataset
