--- src/evaluate/__init__.py
@@ -1,2 +1,0 @@
-from .keras_metrics import KerasMetrics
-from .numpy_metrics import *
