--- src/evaluate/numpy_metrics.py
@@ -1,5 +1,0 @@
-
-import numpy as np
-
-def rmse(y_true, y_pred):
-    return np.sqrt(np.mean((y_true - y_pred)**2))
