--- src/image/__init__.py
@@ -1,1 +1,0 @@
-from .datadirectory import ClassicImageDataDirectory, ImageDataset
