--- src/models/__init__.py
@@ -1,4 +1,0 @@
-from .olsmodel import OLSRegressor
-from .polymodel import PolynomialModel
-
-__all__ = ["OLSRegressor", "PolynomialModel"]
