+++ src/olsmodel.py
@@ -1,7 +1,6 @@
 '''
 OLS Regression model
 '''
-import numpy as np
 
 class OLSRegressor:
