--- src/preprocess/__init__.py
@@ -1,3 +1,0 @@
-from ._transformer_boxcox import BoxCoxTransformer
-
-__all__ = ["BoxCoxTransformer"]
