--- src/tfplay/__init__.py
@@ -1,4 +1,0 @@
-from .tfplaydataset import TfPlayDataset
-from .tfplayground import TensorflowPlayground
-
-__all__ = ["TfPlayDataset", "TensorflowPlayground"]
