--- src/utils/__init__.py
@@ -1,6 +1,0 @@
-from .utils import (describe_tensor, get_dataframe_cols, tensor_variance, 
-                    check_tfmodel_weights_equality)
-from ._labelanalyzer import LabelAnalyzer
-
-__all__ = ["describe_tensor", "get_dataframe_col", "tensor_variance",
-            "check_tfmodel_weights_equality", "LabelAnalyzer"]
