--- src/visualize/__init__.py
@@ -1,7 +1,0 @@
-from .visualize import (plot1d_reg_preds, plot2d_decision_function,
-                        plot_confusion_matrix, plot_keras_model, plot_learning_curve)
-
-
-__all__ = ["plot1d_reg_preds", "plot2d_decision_function",
-           "plot_confusion_matrix", "plot_keras_model", 
-           "plot_learning_curve"]
